assgn1/code/requirements.txt
torch>=1.12
torchvision
numpy
scipy
scikit-learn
matplotlib
tqdm
pillow
opencv-python
scikit-image